Gamer_Recommendation/requirements.txt
fastapi
uvicorn
pydantic
numpy
tensorflow==2.17.0
pandas
scikit-learn
matplotlib
jwt
pyjwt
kafka-python
psycopg2-binary
httpx
google-cloud-secret-manager
apscheduler
networkx
seaborn
requests